+++ event_horizon/api/user/views.py
@@ -1,8 +1,9 @@
 from http import HTTPStatus
 
-from apiflask import APIBlueprint, EmptySchema, pagination_builder
+from apiflask import APIBlueprint, EmptySchema, HTTPError, pagination_builder
+from flask_jwt_extended import jwt_required
 
-from event_horizon.api import PaginationQuery
+from event_horizon.api import PaginationQuery, admin_required
 from event_horizon.api.user.schemas import UserDTO, UserRequestDTO
 from event_horizon.extensions import db
 from event_horizon.models import User
@@ -12,8 +13,10 @@ user_bp = APIBlueprint("users", __name__)
 
 
 @user_bp.get("/users")
+@admin_required()
 @user_bp.input(PaginationQuery, location="query")
 @user_bp.output(UserDTO(many=True))
+@user_bp.doc(security="BearerAuth")
 async def list(query_data):
     paginated_users = db.paginate(
         db.select(User), page=query_data["page"], per_page=query_data["per_page"]
@@ -25,10 +28,14 @@ async def list(query_data):
     }
 
 
-@user_bp.get("/users/<int:id>")
+@user_bp.get("/users/<string:id>")
+@jwt_required()
 @user_bp.output(UserDTO)
+@user_bp.doc(security="BearerAuth")
 async def get(id):
-    user = db.get_or_404(User, id)
+    user = db.session.query(User).filter(User.resource_id == id).first()  # type: ignore
+    if user is None:
+        raise HTTPError(HTTPStatus.NOT_FOUND, "user not found")
     links = (
         generate_links("events", [event.id for event in user.events])  # type: ignore
         if len(user.events) > 0  # type: ignore
@@ -38,8 +45,10 @@ async def get(id):
 
 
 @user_bp.post("/users")
+@admin_required()
 @user_bp.input(UserRequestDTO)
 @user_bp.output(UserDTO, status_code=HTTPStatus.CREATED)
+@user_bp.doc(security="BearerAuth")
 async def create(json_data):
     new_user = User(**json_data)
     db.session.add(new_user)
@@ -47,21 +56,31 @@ async def create(json_data):
     return {"data": new_user}
 
 
-@user_bp.patch("/users/<int:id>")
+@user_bp.patch("/users/<string:id>")
+@jwt_required(fresh=True)
 @user_bp.input(UserRequestDTO(partial=True))
 @user_bp.output(UserDTO)
+@user_bp.doc(security="BearerAuth")
 async def update(id, json_data):
-    user = db.get_or_404(User, id)
+    user = db.session.query(User).filter(User.resource_id == id).first()  # type: ignore
+    if user is None:
+        raise HTTPError(HTTPStatus.NOT_FOUND, "user not found")
+
     for key, value in json_data.items():
         user.__setattr__(key, value)
     db.session.commit()
     return {"data": user}
 
 
-@user_bp.delete("/users/<int:id>")
+@user_bp.delete("/users/<string:id>")
+@admin_required()
 @user_bp.output(EmptySchema, status_code=HTTPStatus.NO_CONTENT)
+@user_bp.doc(security="BearerAuth")
 async def delete(id):
-    user = db.get_or_404(User, id)
+    user = db.session.query(User).filter(User.resource_id == id).first()  # type: ignore
+    if user is None:
+        raise HTTPError(HTTPStatus.NOT_FOUND, "user not found")
+
     db.session.delete(user)
     db.session.commit()
     return None
